+++ src/installers/pip.ts
@@ -7,10 +7,10 @@ export default class PipInstaller extends BaseInstaller {
     async installPackage(name: string, version: string, extras: string[], dev: boolean): Promise<void> {
         const packageName = this.buildPackageName(name, extras);
         const versionConstraint = version ? this.buildVersionConstraint(version) : '';
-        await this.exec(`${this.pythonPath} -m pip install -U "${packageName}${versionConstraint}"`);
+        await this.exec(`${this.mainPythonPath} -m pip install -U "${packageName}${versionConstraint}"`);
     }
 
     async removePackage(name: string, dev: boolean): Promise<void> {
-        await this.exec(`${this.pythonPath} -m pip uninstall -y ${name}`);
+        await this.exec(`${this.mainPythonPath} -m pip uninstall -y ${name}`);
     }
 }