+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to the "boto3-ide" extension will be documented in this file
 ### Fixed
 - Added missing service packages
 - Improved service list update script
+- `poetryPath`/`pipenvPath` now has the highest priority before searching in Python path
+- Skip duplicate Python paths when trying to find installers
 
 ## [0.2.1]
 ### Fixed